Click into the center of the button (text) instead of the corner (#22)
This change is a trail to tackle an issue, that especially the "Sell" button was not hit correcty resulting in the bot being stuck in a loop. Now, the center of the four corner points of the button text is calculated and used as the point to tap for a button click.